requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.1.0
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
crashtest==0.4.1
cryptography==37.0.4
Deprecated==1.2.13
distlib==0.3.6
dulwich==0.20.46
fastapi==0.88.0
filelock==3.8.0
future==0.18.3
googleapis-common-protos==1.56.2
greenlet==2.0.1
grpcio==1.51.1
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
MarkupSafe==2.1.2
more-itertools==8.14.0
msgpack==1.0.4
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-asgi==0.36b0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
opentracing==2.4.0
opentracing-instrumentation==3.3.1
packaging==22.0
pexpect==4.8.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
prometheus-client==0.15.0
prometheus-fastapi-instrumentator==5.9.1
protobuf==3.20.3
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.4
pyrsistent==0.18.1
python-dotenv==0.21.0
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
thrift==0.16.0
tomli==2.0.1
tomlkit==0.11.5
tornado==5.1.1
trove-classifiers==2022.12.22
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
